Improve Physics

+++ App/ChaoticHotel-TvOS/ChaoticHotel-TvOS/Entities/staff/Staff+HoldItens.swift
@@ -12,13 +12,15 @@ import GameplayKit
 
 extension Staff {
     func deliver(entityManager: EntityManager, aService service: ServiceComponent) {
-        print("DELIVER")
         
-        service.deliverService(ofType: self.service)
-        if (self.holdItem != nil && self.service != .clean) {
-            entityManager.remove(self.holdItem)
-            self.holdItem = nil
-            self.service = .listen
+        if (service.waitForServiceType == self.service) {
+            print("DELIVER")
+            service.deliverService(ofType: self.service)
+            if (self.holdItem != nil && self.service != .clean) {
+                entityManager.remove(self.holdItem)
+                self.holdItem = nil
+                self.service = .listen
+            }
         }
     }
 }
